+++ src/esp32/src/api_h/api.h
@@ -10,7 +10,7 @@ void auth_client();
 void revoke_auth_client();
 void revoke_auth();
 void send_data_to_server(uint8_t *data, uint32_t dataLength);
-void send_value_to_server(uint8_t datatType, uint64_t timeStamp, double value);
+void send_value_to_server(uint8_t datatType, uint64_t timeStamp, float value);
 void send_status_to_server(uint64_t timeStamp);
 void send_log_to_server(uint64_t timeStamp, char *level, uint32_t levelLength, char *message, uint32_t messageLength);
