+++ src/esp32/src/storage/storage.h
@@ -15,9 +15,9 @@ typedef struct
     uint8_t identifier[32];
     uint8_t secret[32];
     uint8_t signatureKey[32];
-    double tempTarget;
-    double rpmTarget;
-    double phTarget;
+    float tempTarget;
+    float rpmTarget;
+    float phTarget;
 } preferences_t;
 
 void set_preferences(preferences_t *prefs);
